+++ src/app/app-routing.module.ts
@@ -1,7 +1,22 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-const routes: Routes = [];
+import { HomePageComponent } from './pages/home-page/home-page.component';
+import { authGuard } from './guards/auth.guard';
+
+const routes: Routes = [
+  {
+    path: 'home',
+    pathMatch: 'full',
+    component: HomePageComponent,
+    canActivate: [authGuard],
+  },
+  {
+    path: '',
+    pathMatch: 'full',
+    redirectTo: 'home',
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],