+++ src/app/services/modals/app-modals.model.ts
@@ -1,0 +1,3 @@
+import { DialogConfig } from '@angular/cdk/dialog';
+
+export type AppModalConfig = Omit<DialogConfig, 'data' | 'providers' | 'container'>;